Avoid errors when compiling with clang
This commit fixes all the compilation errors that are triggered by the
option "-Werror=missing-prototypes", when compiling with clang instead
of gcc in Debian GN/Linux .  Most of the errors were caused by
functions that are locally defined but were lacking the "static"
specifier.  This is fixed by this commit.

Function declarations (either with or without the "extern" specifier)
appearing in *.cpp files have been deleted and moved into appropriate
*.h header files.  I hope I got them all right.

Some functions, which were defined but used nowhere, are removed.
They are:

  - CC_getValueAtTime and CC_getC0ValueAtTime (dwtools/CC.cpp)
  - Minimizer_setAfterEachIteration (dwtools/Minimizers.cpp)
  - espeak_SetPhonemeCallback (external/espeak/speak_lib.cpp)
  - IsAllUpper (external/espeak/translate.cpp)

Note that a similar patch [1] has been proposed for the Debian praat
package version 6.0.5-1, which closed Bug#812715 [2].  However, that
patch simply adds a function declaration before each function
definition. The present commit addresses the issue in a more elegant
way, I hope.

[1] https://anonscm.debian.org/cgit/debian-med/praat.git/commit/?id=b3356c3
[2] https://bugs.debian.org/812715