No return value for function init_genrand64
The return value of function init_genrand64 seems to be useless.  It
appears to be a leftover from a past change that has no function now.
Indeed, init_genrand64 is called twice in the code, once in
NUMrandom_State::init_by_array64, which makes no use of the return
value, and once in NUMrandom_initializeWithSeedUnsafelyButPredictably.
In the later, the value returned from init_genrand64 is used in a
strange way, put into the input argument "seed".  This commit also
fixes this.