FreeBSD: introduce a makefile.defs.freebsd.alsa
This is the same makefile.defs as is used in packaging downstream.
Various settings are picked up in the packaging system from
"outside" which might need to be set when building by hand;
hence ?= assignments.

The CPPFLAGS are not added automatically to builds, so those
are attached to the CFLAGS and CXXFLAGS (so they do pick up
settings in the sub-Makefiles).