Makefiles: split external build from Praat-itself
This introduces intermediate targets all-external and clean-external
that deal with the bundled external packages, and also all-self and
clean-self that deal with the Praat sources themselves.

This is inspired by #1238, and makes it slightly easier to
build externals while ignoring warnings, and then building
Praat itself while paying attention to warnings.