Replace "three valued" bool with an enum
Since *garnish* was a bool, it could sensibly only have two
values (false, true), but it was being used as an int,
with values 0, 1, 2 being passed in and tested-for.

Replace with an enum class, so it's more clear what the
values actually mean; since this is an internal, static method
the scope is quite restricted.

(Note: I've given the enum values names based on my understanding
of what the code is drawing here; there are probably better names)